+++ CHANGELOG.md
@@ -30,8 +30,8 @@
 - [node] feat: Make node transactions a pluggable integration with tests
 - [node] feat: Transactions handling for RequestHandler in Express/Hapi
 - [node] fix: Dont wrap native modules more than once to prevent leaks
+- [node] fix: Add the same protocol as dsn to base transport option
 - [node] fix: Use getCurrentHub to retrieve correct hub in requestHandler
-
 - [utils] ref: implemented includes, assign and isNaN polyfills
 
 ## 4.0.6